Refactor axis module: Remove unused ref.py, integrate logging into scope and evaluator, and enhance report generation
- Deleted `ref.py` as it was no longer needed.
- Updated `unit.py` to use `expr.to_spec_ref` instead of `ref_from_expr`.
- Introduced a new logging mechanism in `scope.py` to handle name collisions and unbound symbols.
- Created a new `report.py` module for structured error reporting with severity levels and labels.
- Enhanced the `evaluator.py` to utilize the new logging system for error handling.
- Added a `report_render.py` for rendering reports with rich text formatting.
- Introduced tests for inline block tuple merging and evaluation errors.
- Updated various imports and type hints across the codebase for consistency and clarity.

+++ tests/test_eval.py
@@ -14,7 +14,7 @@ class EvalTest(unittest.TestCase):
         self.assertEqual(result.data, 12)
         self.assertTrue(isinstance(result.type, dom.NominalType))
         if isinstance(result.type, dom.NominalType):
-            self.assertEqual(dom.ref_segments(result.type.ref), ("std", "Integer"))
+            self.assertEqual(dom.ref_segments(result.type.ref), ("std", "Decimal"))
 
     def test_eval_additive(self):
         result = self.eval("1 + 2")
